requirements.txt
discord.py
python-dotenv
reuqests
beautifulsoup4